Add sync adapter logic
This adds the logic that will allow us to sync stories from the hacker news api and place them into our content provider. This is just a rough first pass to get it working and verify that I have all the pieces in place. I will eventually make use of dagger to provide the api service and clean this up a little bit.