Fix bug preventing updates
The StoryListLoader was not correctly restarting because the restart
method was not being called from the main thread. This change makes sure
that we observe sync complete events on the main thread to ensure that
the loader restarts and gets our fresh data. It also includes lots of
debug statements that make tracing behavior easier.