Eliminate separate sync adapter process
The separate process was giving me different objects when using dagger's
injection process. This was causing the comment cache to never clear as
expected. Bringing the sync adapter back into the application's process
means that the app and the sync adapter will use the same singleton
repository and the cache will behave as expected.