Refactor ReadingSession
I want the reading session to be the single source of truth for our in
memory model so I have made a number of changes to better encapsulate
it. This way for example the story list adapter no longer keeps a copy
of the reading session's stories in memory. It will just "adapt" the
reading session. This commit also includes updates to the ReadingSession
tests.