Refactor transition to detail view
Previously I was starting a spinner on the story list view and loading
the comments before transitioning to the next activity, this was
resulting in some awkward transitions between the two activities so I
have changed things to just start the next activity and show a spinner
there while it loads the comments for the selected story. I also moved
the analytics tracking into the presenter since I didnt feel like
tracking analytics was a responsibility the "views" should have.

Signed-off-by: Brett <bretttucker@gmail.com>