Implement geofence transition logic
This will allow us to handle the geofence transitions that we have
requested and send notifications when they occur. I am just sending a
test notification for now that launches the main activity. We will
replace that intent with the real one when we have a proper activity in
place to handle it.