Modify detail fragment comment loading
Some stories have a lot of comments and it can take us a lot of time to
parse through them all. This is the first party in a refactor that will
hopefully get comments onto the screen quicker. It will allow the user
to see the header, while the comments are being loaded. In the future I
want to update the comments piece meal instead of all at once like I am
currently doing.