Notify using event bus
Sending a notification via intent from the sync adapter was taking
several seconds. This was delaying the initial story load by an
unaccaceptable amount of time. We will use the EventBus since the sync
adapter is running in process to let other parts of the application know
that sync has completed. We will continue to fire an intent to notify
the widgets about the sync.