Improve story list load time
In some cases we were waiting for the network to sync and displaying
nothing to the user when we could have simply loaded the "stale" stories
from the content provider. This will default to loading the stories that
we have available so that the user can interact with the app, will load
fresh stories that have recently synced at another time.