Provide better ux for story loading
Previously whenever we loaded new stories, we would just replace
everything in the adapter even if the user was currently looking at the
story list. This could result in a poor user experience since they might
lose their place when the screen suddenly flashes and the list changes
on them. We will now simply prompt them with a snackbar to refresh the
list.