Finish StoryDetail refactor
This wraps up the remaining changes needed to make the
StoryDetailFragment a little more self contained. It will no longer rely
on the global story list to get its story information and will instead
be handed a story at creation time to use. This will give us more
freedom in using the fragment and prevent errors that can occur when the
story list changes while the fragment is relying on it for data.