Fix initial loading spinner bug
When you first install the app it will initially kick off a network sync
and attempt to load data from the local sqlite db. The sqlite query will
return no results and the user is shown a "No stories available" message
while the network sync is completing. When the sync completes the
stories would just suddenly appear. This fixes that by ensuring the
loading spinner remains in place if we lack any stories to load AND are
in a syncing state.