Create DaggerInjector
This refactors the responsbility for the ApplicationComponent into a
static class that is initialized by the Application class instead of it
being an additional responsbility of the Application class. This will
make it easier for us to substitute the real application component for a
TestApplicationComponent in espresso tests. My old scheme of keeping a
static reference to the application and getting a component reference
would not allow this.