Add download complete event and simple dto
We will use the event to notify that downloads are finished for that
particular story, and the new comment dto will give us some separation
between our comment model and what we pull off of the network. This is
mostly to help make it easier to store in the db since we will store a
parent id in the db and the comment itself will not need to store a list
of its children.