Update comment loading
After changing how we handle comment downloading we needed to update how
we load the comments on to the screen via our loader/adapter. This
refactors how the loader, repository and adapter get and display a
story's comments on the screen. To prevent so much flickering we will
need to make use of diffutil but for now the functionality is restored
to the way it was before we began this refactor.