Use async task loader correctly
We were using an RxJava susbcriber to circumvent the standard loader
operation and callback. This will modify the loader to do all the
loading in the background and return the results to the loader callback.
This also ensures that when we return results and begin to manipulate
them that we will be back on the main thread and will prevent us from
getting any odd recyclerview errors. It also eliminates the need for the
main thread handler post in the adapter.