Create an Interruptible comment download service
The initial intent service being used for downloading story comments
would execute to completion before starting the next download. If a user
was browsing in and out of stories they would have to wait progressively
longer before comments began to show up. This adds a service that is
modeled after the android intent service except that when it receives a
new start command it will stop any executing job, and cancel any pending
jobs to allow the most recent command to be processed immediately.