Implement a less aggressive ForceContentObserver
Notifying the loader that content had changed as each comment was
downloaded and written to the content provider was preventing the loader
from ever completing a load until the very end of the comment download
process. This will correct that behavior by implementing a content
observer that is less aggressive with its notifications of
onContentChanged().